Working on moving movie scanner to its own package
