+++ server/cmd/api/scanner.go
@@ -2,18 +2,13 @@ package main
 
 import (
 	"context"
-	"igloo/cmd/internal/helpers"
+	"igloo/cmd/internal/scanner"
 )
 
-// scannerBatchSize is how many files a library scan buffers before flushing them
-// to the database.
-const scannerBatchSize = 54
+const scannerBatchSize = scanner.BatchSize
 
-// Library scans are single-flighted across the startup scan and the
-// admin-triggered scan endpoints.
 var (
-	movieScanGuard helpers.ScanGuard
-	musicScanGuard helpers.ScanGuard
+	musicScanGuard scanner.ScanGuard
 )
 
 // scanContext returns the shutdown-aware context library scans run under, and
