+++ server/cmd/api/scanner_test_fixtures_test.go
@@ -1,0 +1,91 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"strings"
+	"testing"
+
+	"igloo/cmd/internal/database"
+	"igloo/cmd/internal/tmdb"
+)
+
+// noKeyframeProbe completes ffprobe.FfprobeInterface for scanner stubs that
+// never serve HLS.
+type noKeyframeProbe struct{}
+
+func (noKeyframeProbe) KeyframeAtOrBefore(context.Context, string, int64, float64) (float64, error) {
+	return 0, errors.New("keyframe probing is not stubbed")
+}
+
+type stubMovieScannerTmdb struct {
+	searchErr     error
+	detailErr     error
+	theatersErr   error
+	searchResults []tmdb.TmdbMovie
+	detailMovies  map[int]tmdb.TmdbMovie
+	theaterMovies []*tmdb.TmdbMovie
+	searchCalls   []stubMovieScannerTmdbSearchCall
+	detailCalls   []int
+}
+
+type stubMovieScannerTmdbSearchCall struct {
+	title string
+	year  []int
+}
+
+func (s *stubMovieScannerTmdb) GetTmdbMovieByID(_ context.Context, movie *tmdb.TmdbMovie) error {
+	s.detailCalls = append(s.detailCalls, movie.TmdbID)
+	if s.detailErr != nil {
+		return s.detailErr
+	}
+	if s.detailMovies == nil {
+		return errors.New("tmdb details unavailable")
+	}
+	details, ok := s.detailMovies[movie.TmdbID]
+	if !ok {
+		return errors.New("tmdb details unavailable")
+	}
+	*movie = details
+	return nil
+}
+
+func (s *stubMovieScannerTmdb) SearchMoviesByTitleAndYear(_ context.Context, title string, year ...int) ([]tmdb.TmdbMovie, error) {
+	yearCopy := append([]int(nil), year...)
+	s.searchCalls = append(s.searchCalls, stubMovieScannerTmdbSearchCall{title: title, year: yearCopy})
+	if s.searchErr != nil {
+		return nil, s.searchErr
+	}
+	results := make([]tmdb.TmdbMovie, len(s.searchResults))
+	copy(results, s.searchResults)
+	return results, nil
+}
+
+func (s *stubMovieScannerTmdb) GetMoviesInTheaters(_ context.Context) ([]*tmdb.TmdbMovie, error) {
+	if s.theatersErr != nil {
+		return nil, s.theatersErr
+	}
+	return s.theaterMovies, nil
+}
+
+func (*stubMovieScannerTmdb) ClearCache() {}
+
+func tmdbMovieFromJSON(t *testing.T, payload string) tmdb.TmdbMovie {
+	t.Helper()
+
+	var movie tmdb.TmdbMovie
+	err := json.Unmarshal([]byte(payload), &movie)
+	if err != nil {
+		t.Fatalf("unmarshal TMDB movie fixture: %v", err)
+	}
+	return movie
+}
+
+func movieGenreTags(genres []database.GetGenresByMovieIDRow) string {
+	tags := make([]string, 0, len(genres))
+	for _, genre := range genres {
+		tags = append(tags, genre.Tag)
+	}
+	return strings.Join(tags, ",")
+}
