+++ server/cmd/internal/scanner/scanner_test.go
@@ -1,4 +1,4 @@
-package helpers
+package scanner
 
 import (
 	"context"
